requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-flask>=1.2.0
pytest-xdist>=3.3.0

# Mocking and fixtures
responses>=0.23.0
requests-mock>=1.11.0
factory-boy>=3.3.0

# Test utilities
freezegun>=1.2.0
testfixtures>=7.2.0
faker>=19.0.0

# Load testing
locust>=2.16.0